Cast topology pointers as ints

+++ shapely/vectorized/vectorfuncs.pyx
@@ -119,12 +119,12 @@ class BinaryRealPropertyVec(DelegatingVec):
 class UnaryTopologicalOpVec(UnaryVec):
 
     def __call__(self, this, *args):
-        return self.fn(this)
+        return self.fn(this).astype(int)
 
 class BinaryTopologicalOpVec(BinaryVec):
 
     def __call__(self, this, others, *args):
-        return self.fn(this, others)
+        return self.fn(this, others).astype(int)
